feat：Move `FeedImageDataLoaderPresentationAdapter` to separate file
